Improve requirements handling (#1287)
* setup: read requirements.txt for dependencies

This makes it easier to keep required packages in sync.

Signed-off-by: Paul Spooren <mail@aparcar.org>

* requirements: Update to click 5.0 and redis 3.5.0

Click 5.0 was already required by the `setup.py` and is not brought in
sync. Redis Python library 3.5.0 introduces the `HSET` command with
mapping support which replaces the previous `hmset`. By lifting the
minimal required version to 3.5.0 we can remove the combat function if
Redis server 4.0 is guaranteed.

Signed-off-by: Paul Spooren <mail@aparcar.org>

* ci: remove Python3.4 testing

`redis-py` 3.5.0 does no longer support Python 3.4, so drop it in CI.

Signed-off-by: Paul Spooren <mail@aparcar.org>