Add ssl_cert_reqs to cli as proposed in #1468
Make REDIS_SSL_CERT_REQS available as a config/settings option. This is to be able to set this state from the default required to None, what is needed to deploy with SSL = True but without proper certificates (e.g. within datacenter like Heroku). Pull request proposed by @mgcdanny in #1468